+++ PS7-x64/Modules/AdminToolbox/7.1.0.0/Public/Get-Info.ps1
@@ -12,6 +12,7 @@ Get-All                           ..Gets all module Functions
 Get-Full                          ..Gets full help for selected module command
 Get-ModuleAliases                 ..Gets aliases for AdminToolbox modules
 Invoke-Show                       ..Invokes a grid Gui for implementing show command
+Open-AdminMMC                     ..Opens Microsoft Management Console with multiple snap-ins loaded
 Start-Application                 ..Start an application from PowerShell with ease
           "                                                                                                 -ForegroundColor cyan
     Write-Host " "