+++ PS7-x64/Modules/AdminToolbox.Exchange/1.2.0.0/ChangeLog.md
@@ -8,3 +8,4 @@
 * **1.0.2.0** Added markdown to the changelog
 * **1.0.3.0** Spellcheck Everything
 * **1.1.0.0** Tested and marked Core compatbile
+* **1.2.0.0** Added 2 functions