+++ PS7-x64/Modules/AdminToolbox/7.1.0.0/ChangeLog.md
@@ -123,3 +123,4 @@
 * **7.0.0.0**  Added BetterCredentials 3rd party module requirement, and added Get-Filename and Get-Foldername Functions.
 * **7.0.0.1**  Remove Bettercredentials Required modules
 * **7.0.1.1**  Updated get-info to reflect changes in the last update
+* **7.1.0.0**  Added function Open-AdminMMC